+++ datumaro/components/extractor.py
@@ -16,6 +16,9 @@ from datumaro.components.annotation import (
 )
 from datumaro.components.cli_plugin import CliPlugin
 from datumaro.components.errors import DatasetNotFoundError
+from datumaro.components.format_detection import (
+    FormatDetectionConfidence, FormatDetectionContext,
+)
 from datumaro.util import is_method_redefined
 from datumaro.util.attrs_util import default_if_none, not_empty
 from datumaro.util.image import Image
@@ -209,11 +212,13 @@ class SourceExtractor(Extractor):
 
 class Importer(CliPlugin):
     @classmethod
-    def detect(cls, path):
-        if not path or not osp.exists(path):
-            return False
+    def detect(
+        cls, context: FormatDetectionContext,
+    ) -> Optional[FormatDetectionConfidence]:
+        if not cls.find_sources_with_params(context.root_path):
+            context.fail("specific requirement information unavailable")
 
-        return len(cls.find_sources_with_params(path)) != 0
+        return FormatDetectionConfidence.LOW
 
     @classmethod
     def find_sources(cls, path) -> List[Dict]: