+++ tests/test_mots_format.py
@@ -6,6 +6,7 @@ import numpy as np
 
 from datumaro.components.annotation import Mask
 from datumaro.components.dataset import Dataset
+from datumaro.components.environment import Environment
 from datumaro.components.extractor import DatasetItem
 from datumaro.plugins.mots_format import MotsImporter, MotsPngConverter
 from datumaro.util.image import Image
@@ -130,7 +131,8 @@ class MotsPngConverterTest(TestCase):
 class MotsImporterTest(TestCase):
     @mark_requirement(Requirements.DATUM_GENERAL_REQ)
     def test_can_detect(self):
-        self.assertTrue(MotsImporter.detect(DUMMY_DATASET_DIR))
+        detected_formats = Environment().detect_dataset(DUMMY_DATASET_DIR)
+        self.assertIn(MotsImporter.NAME, detected_formats)
 
     @mark_requirement(Requirements.DATUM_GENERAL_REQ)
     def test_can_import(self):